+++ graph.h
@@ -3,6 +3,7 @@
 #include <string>
 #include <iostream>
 #include "vertex.h"
+
 #define VERTICES 5
 
 struct Graph
@@ -11,8 +12,8 @@ struct Graph
     int size = 0;
 };
 
-void connectVertex(Vertex*& vertex, int connectedTo, int distance);
-Graph createGraphFromMatrix(const int matrix[][VERTICES], int n);
+void connectVertex(Vertex*& vertex, int connectedTo, float distance);
+Graph createGraphFromMatrix(const float matrix[][VERTICES], int n);
 void deleteGraph(Graph& graph);
 void deleteVertices(Vertex* vertex);
 void printGraph(const Graph& graph);
